--- code.py
@@ -1,4 +1,0 @@
-from main import read_from_input, display_logo
-
-display_logo("mister")
-read_from_input()
